KH_Potential_Evaluation/requirements.txt
fastapi
logging
openai
pydantic
PyYAML
uvicorn